Keep consts together; move `getPolicyRegistry`
