+++ src/utils/getJoinPolicyRegistry.ts
@@ -1,9 +1,9 @@
 import { Contract, providers } from 'ethers'
 import { Provider } from '@web3-react/types'
-import { JoinPolicyRegistryAddress } from '$/features/tokenGatedRooms/utils/const'
+import { JoinPolicyRegistryAddress } from '$/consts'
 import { abi } from '$/contracts/JoinPolicyRegistry.sol/JoinPolicyRegistry.json'
 
-export default function getPolicyRegistry(provider: Provider) {
+export default function getJoinPolicyRegistry(provider: Provider) {
     return new Contract(
         JoinPolicyRegistryAddress,
         abi,
